Introduce manual timestamp posting
After having moved to the built-in timestamp for conditions on the smart contracts, we must still be able to keep posting timestamps the next 3 months to make sure all previously issue smart contracts have a timestamp that would fulfil the condition. After those three months, the manual posting can be removed.